test: add transcript integration tests and fix equals() numeric comparison
- Add 4 integration tests in cmd_test.go
- Fix equals() to handle cross-type numeric comparison (int vs float64)
- Change builtinTranscriptCount to return float64 for type consistency

+++ internal/expression/evaluator_test.go
@@ -1704,7 +1704,7 @@ func TestTranscriptCount(t *testing.T) {
 		t.Fatalf("error: %v", err)
 	}
 
-	if result != 2 {
+	if result != float64(2) {
 		t.Errorf("expected count=2, got %v", result)
 	}
 }
@@ -1719,7 +1719,7 @@ func TestTranscriptCount_NoMatches(t *testing.T) {
 		t.Fatalf("error: %v", err)
 	}
 
-	if result != 0 {
+	if result != float64(0) {
 		t.Errorf("expected count=0, got %v", result)
 	}
 }
